Move components and add raw unit test for website component.
