+++ lib/components/roots/ThemeProvider/index.tsx
@@ -1,6 +1,6 @@
 import { ThemeProvider as JssThemeProvider } from 'react-jss';
-import { Theme } from '../..';
-import { themes } from '../../themes';
+import { Theme } from '../../..';
+import { themes } from '../../../themes';
 import React, { FunctionComponent, ReactElement } from 'react';
 
 interface ThemeProviderProps {
