+++ lib/components/roots/Website/index.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
-import { createUseStyles } from '../../styles';
-import { Dialogs } from '../Dialogs';
-import { Icons } from '../graphics/Icons';
-import { Notifications } from '../Notifications';
-import { Theme } from '../../themes';
+import { createUseStyles } from '../../../styles';
+import { Dialogs } from '../../Dialogs';
+import { Icons } from '../../graphics/Icons';
+import { Notifications } from '../../Notifications';
+import { Theme } from '../../../themes';
 import React, { CSSProperties, FunctionComponent, ReactElement } from 'react';
 import { styles, WebsiteClassNames } from './styles';
