+++ lib/index.ts
@@ -30,15 +30,15 @@ import { Sidebar } from './components/layout/Sidebar';
 import { SidebarBrand } from './components/layout/SidebarBrand';
 import { SidebarFooter } from './components/layout/SidebarFooter';
 import { SidebarItem } from './components/layout/SidebarItem';
-import { StyleCollector } from './components/StyleCollector';
+import { StyleCollector } from './components/roots/StyleCollector';
 import { Text } from './components/typography/Text';
 import { TextArea } from './components/input/TextArea';
 import { TextBox } from './components/input/TextBox';
-import { ThemeProvider } from './components/ThemeProvider';
+import { ThemeProvider } from './components/roots/ThemeProvider';
 import { Toggle } from './components/Toggle';
 import { Transition } from './components/animation/Transition';
 import { TransitionGroup } from './components/animation/TransitionGroup';
-import { Website } from './components/Website';
+import { Website } from './components/roots/Website';
 import { classNames, createUseStyles, removeServerSideStyles, ServerSideStyles, useTheme } from './styles';
 import { dialogs, excecutionEnvironment, getPortalRootNode, notifications } from './services';
 import { Form, FormActions, FormRow } from './components/Form';
