+++ test/unit/roots/WebsiteTests.tsx
@@ -1,0 +1,33 @@
+import { act } from '../../shared/act';
+import { assert } from 'assertthat';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { ThemeProvider, Website } from '../../../lib';
+
+suite('Website', (): void => {
+  let container: Element;
+
+  setup((): void => {
+    container = document.createElement('div');
+    document.body.appendChild(container);
+  });
+
+  teardown((): void => {
+    document.body.removeChild(container);
+  });
+
+  test('renders.', async (): Promise<void> => {
+    act((): void => {
+      ReactDOM.render(
+        <ThemeProvider>
+          <Website>Website</Website>
+        </ThemeProvider>,
+        container
+      );
+    });
+
+    const website = container.querySelector('div');
+console.log(website)
+    assert.that(website!.className).is.containing('Website');
+  });
+});
